Made mqtt-bridge pythonic. Added linting. Added poetry package management
